[TOW-102] 휴식 타이머 수정
* Present 방식에서 ZStack 방식으로 변경
* 0초 후 재시작 처리 수정
* BreakTimerFeature 제거 후 뷰로 변경
