+++ TodayWod/Features/Home/WorkOut/Detail/WorkOutDetailFeature.swift
@@ -17,8 +17,10 @@ struct WorkOutDetailFeature {
         var duration: Int
         var hasStart: Bool
         var isDayCompleted: Bool
-        
-        @Presents var timerState: BreakTimeFeature.State?
+
+        var currentSeconds: Int = 0
+        var buttonState: ButtonState = .pause
+
         @Presents var confirmState: WorkoutConfirmationFeature.State?
         
         init(item: DayWorkoutModel) {
@@ -43,14 +45,27 @@ struct WorkOutDetailFeature {
         case updateWodSet(WodSetModel)
         case saveOwnProgram
         case updateDayCompleted
-        case breakTimerAction(PresentationAction<BreakTimeFeature.Action>)
         case binding(BindingAction<State>)
         case confirmAction(PresentationAction<WorkoutConfirmationFeature.Action>)
         case finishWorkOut(DayWorkoutModel)
+
+        case breakTimerTick
+        case didTapReset
+        case didTapPause
+        case didTapResume
+        case setButtonState
     }
-    
-    enum CancelID { case timer }
-    
+
+    enum ButtonState {
+        case pause
+        case resume
+    }
+
+    enum CancelID {
+        case timer
+        case breakTimer
+    }
+
     @Dependency(\.dismiss) var dismiss
     
     var body: some ReducerOf<Self> {
@@ -90,13 +105,13 @@ struct WorkOutDetailFeature {
             case let .setCompleted(set):
                 var updatedSet = set
                 updatedSet.isCompleted.toggle()
-                
-                if updatedSet.isCompleted {
-                    state.timerState = BreakTimeFeature.State()
-                }
-                
                 let wodSet = updatedSet
-                return .run { send in
+
+                return updatedSet.isCompleted
+                ? .merge(.run { send in
+                    await send(.updateWodSet(wodSet))
+                }, .send(.didTapReset))
+                : .run { send in
                     await send(.updateWodSet(wodSet))
                 }
             case let .setUnitText(unitText, set):
@@ -113,8 +128,8 @@ struct WorkOutDetailFeature {
                         if let setIndex = wod.wodSets.firstIndex(where: { $0.id == set.id }) {
                             state.item.workouts[index].wods[wodIndex].wodSets[setIndex] = set
                             
-                            print("%%%%%% UpdateWodSet: \(state.item.workouts)")
-                            
+                            DLog.d("%%%%%% UpdateWodSet: \(state.item.workouts)")
+
                             break outerLoop
                         }
                     }
@@ -135,7 +150,9 @@ struct WorkOutDetailFeature {
                 
                 if state.isDayCompleted {
                     state.confirmState = WorkoutConfirmationFeature.State(type: .completed) // 운동 완료 재확인.
-                    state.timerState = BreakTimeFeature.State()
+
+                    // TODO: - 이건 왜 하는지 확인 필요
+                    // state.timerState = BreakTimeFeature.State()
                 }
                 return .none
             case .confirmAction(.presented(.didTapDoneButton)):
@@ -146,12 +163,51 @@ struct WorkOutDetailFeature {
                 return .concatenate(.send(.stopTimer),
                                     .send(.saveOwnProgram),
                                     .send(.finishWorkOut(state.item)))
-            default:
+            case .binding:
                 return .none
+            case .confirmAction:
+                return .none
+            case .finishWorkOut:
+                return .none
+            case .breakTimerTick:
+                if state.currentSeconds <= 0 {
+                    return .cancel(id: CancelID.breakTimer)
+                }
+                state.currentSeconds -= 1
+                return .none
+            case .didTapReset:
+                state.currentSeconds = 60
+                state.buttonState = .pause
+                return .concatenate(
+                    .cancel(id: CancelID.breakTimer),
+                    .run { send in
+                        while true {
+                            try await Task.sleep(for: .seconds(1))
+                            await send(.breakTimerTick)
+                        }
+                    }
+                    .cancellable(id: CancelID.breakTimer)
+                )
+            case .didTapPause:
+                return .cancel(id: CancelID.breakTimer)
+            case .didTapResume:
+                return .run { send in
+                    while true {
+                        try await Task.sleep(for: .seconds(1))
+                        await send(.breakTimerTick)
+                    }
+                }
+                .cancellable(id: CancelID.breakTimer)
+            case .setButtonState:
+                switch state.buttonState {
+                case .pause:
+                    state.buttonState = .resume
+                    return .send(.didTapPause)
+                case .resume:
+                    state.buttonState = .pause
+                    return .send(.didTapResume)
+                }
             }
-        }
-        .ifLet(\.$timerState, action: \.breakTimerAction) {
-            BreakTimeFeature()
         }
         .ifLet(\.$confirmState, action: \.confirmAction) {
             WorkoutConfirmationFeature()
@@ -210,12 +266,11 @@ struct WorkOutDetailView: View {
                     }
                     .padding(.horizontal, 38)
                     .padding(.bottom, 20)
+                } else {
+                    BreakTimerView(store: store)
                 }
             }
             .toolbar(.hidden, for: .navigationBar)
-            .bottomSheet(item: $store.scope(state: \.timerState, action: \.breakTimerAction)) { breakStore in
-                BreakTimerView(store: breakStore)
-            }
             .sheet(item: $store.scope(state: \.confirmState, action: \.confirmAction)) { store in
                 WorkoutConfirmationView(store: store)
                     .presentationDetents([.height(dynamicHeight + 20.0)])
